--- EnemyFabricInterface.java
@@ -1,16 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Interface.java to edit this template
- */
-package mortalkombatbversion;
-
-import javax.swing.JLabel;
-import javax.swing.JProgressBar;
-
-/**
- *
- * @author Мария
- */
-public interface EnemyFabricInterface {
-    public Player create(int i);
-}
