+++ LabaLegacy/src/main/java/mortalkombatbversion/CharacterAction.java
@@ -4,6 +4,12 @@
  */
 package mortalkombatbversion;
 
+import mortalkombatbversion.Characters.LiuKang;
+import mortalkombatbversion.Characters.ShaoKahn;
+import mortalkombatbversion.Characters.SonyaBlade;
+import mortalkombatbversion.Characters.SubZero;
+import mortalkombatbversion.Characters.Baraka;
+import mortalkombatbversion.Characters.EnemyFabric;
 import javax.swing.ImageIcon;
 import javax.swing.JDialog;
 import javax.swing.JLabel;
