+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/Baraka.java
@@ -1,0 +1,16 @@
+package mortalkombatbversion.Characters;
+
+import mortalkombatbversion.Player;
+
+public class Baraka extends Player{
+    
+    public Baraka(int level, int health, int  damage, int attack){
+        super (level, health, damage, attack);
+    }
+    
+    @Override
+    public String getName(){
+        return "Baraka";
+    }
+    
+}
