+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/BarakaFabric.java
@@ -1,8 +1,7 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
-package mortalkombatbversion;
+package mortalkombatbversion.Characters;
+
+import mortalkombatbversion.Characters.EnemyFabricInterface;
+import mortalkombatbversion.Player;
