+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/EnemyFabricInterface.java
@@ -1,0 +1,9 @@
+package mortalkombatbversion.Characters;
+
+import javax.swing.JLabel;
+import javax.swing.JProgressBar;
+import mortalkombatbversion.Player;
+
+public interface EnemyFabricInterface {
+    public Player create(int i);
+}
