+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/LiuKang.java
@@ -1,0 +1,15 @@
+package mortalkombatbversion.Characters;
+
+import mortalkombatbversion.Player;
+
+public class LiuKang extends Player{
+    
+    public LiuKang(int level, int health, int  damage, int attack){
+        super (level, health, damage, attack);
+    }
+    
+    @Override
+    public String getName(){
+        return "Liu Kang";
+    }
+}
