+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/LiuKangFabric.java
@@ -1,0 +1,14 @@
+package mortalkombatbversion.Characters;
+
+import mortalkombatbversion.Characters.EnemyFabricInterface;
+import mortalkombatbversion.Player;
+
+public class LiuKangFabric implements EnemyFabricInterface {
+
+    @Override
+    public Player create(int i) {
+        Player enemy;
+        enemy = new LiuKang(1, 70, 20, 1);
+        return enemy;
+    }
+}
