+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/ShaoKahnFabric.java
@@ -2,7 +2,10 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package mortalkombatbversion;
+package mortalkombatbversion.Characters;
+
+import mortalkombatbversion.Characters.EnemyFabricInterface;
+import mortalkombatbversion.Player;
 
 /**
  *
