+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/SonyaBlade.java
@@ -2,20 +2,22 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package mortalkombatbversion;
+package mortalkombatbversion.Characters;
+
+import mortalkombatbversion.Player;
 
 /**
  *
  * @author Мария
  */
-public class LiuKang extends Player{
+public class SonyaBlade extends Player{
     
-    public LiuKang(int level, int health, int  damage, int attack){
+    public SonyaBlade (int level, int health, int  damage, int attack){
         super (level, health, damage, attack);
     }
     
     @Override
     public String getName(){
-        return "Liu Kang";
+        return "Sonya Blade";
     }
 }
