+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/SonyaBladeFabric.java
@@ -2,20 +2,23 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package mortalkombatbversion;
+package mortalkombatbversion.Characters;
 
+import mortalkombatbversion.Characters.EnemyFabricInterface;
+import mortalkombatbversion.Player;
 
 
 /**
  *
  * @author Мария
  */
-public class LiuKangFabric implements EnemyFabricInterface {
+public class SonyaBladeFabric implements EnemyFabricInterface {
 
     @Override
     public Player create(int i) {
         Player enemy;
-        enemy = new LiuKang(1, 70, 20, 1);
+        enemy = new SonyaBlade(1, 80, 16, 1);
         return enemy;
     }
+
 }
