+++ LabaLegacy/src/main/java/mortalkombatbversion/Characters/SubZero.java
@@ -2,20 +2,22 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package mortalkombatbversion;
+package mortalkombatbversion.Characters;
+
+import mortalkombatbversion.Player;
 
 /**
  *
  * @author Мария
  */
-public class ShaoKahn extends Player{
+public class SubZero extends Player{
     
-    public ShaoKahn(int level, int health, int  damage, int attack){
+    public SubZero(int level, int health, int damage , int attack){
         super (level, health, damage, attack);
     }
     
     @Override
     public String getName(){
-        return "Shao Kahn";
+        return "Sub-Zero";
     }
 }
