+++ LabaLegacy/src/main/java/mortalkombatbversion/Human.java
